+++ src/components/compositing/compositing.rs
@@ -50,6 +50,7 @@ pub use constellation::Constellation;
 pub mod compositor_task;
 
 mod compositor_data;
+mod events;
 
 mod compositor;
 mod headless;
